ZQL Refactor
Improving backend design, setting up for arg less than/greater than and other functional primitives like representative
