+++ src/main/java/edu/uiuc/zenvisage/zql/AxisVariable.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package edu.uiuc.zenvisage.zqlcomplete.querygraph;
+package edu.uiuc.zenvisage.zql;
 
 import java.util.ArrayList;
 import java.util.List;
