+++ src/main/java/edu/uiuc/zenvisage/zql/ScatterVCNode.java
@@ -1,4 +1,4 @@
-package edu.uiuc.zenvisage.zqlcomplete.querygraph;
+package edu.uiuc.zenvisage.zql;
 
 import java.io.IOException;
 import java.sql.SQLException;
