+++ src/main/java/edu/uiuc/zenvisage/zql/VisualComponent1.java
@@ -1,4 +1,4 @@
-package edu.uiuc.zenvisage.zqlcomplete.querygraph;
+package edu.uiuc.zenvisage.zql;
 
 import java.util.ArrayList;
 import java.util.List;
