+++ src/main/java/edu/uiuc/zenvisage/zql/functions/D.java
