+++ src/main/java/edu/uiuc/zenvisage/zql/functions/DEuclidean.java
