+++ src/main/java/edu/uiuc/zenvisage/zql/functions/R.java
@@ -1,0 +1,8 @@
+package edu.uiuc.zenvisage.zql.functions;
+
+import edu.uiuc.zenvisage.data.remotedb.VisualComponentList;
+import edu.uiuc.zenvisage.zql.AxisVariableScores;
+
+public interface R {
+	public AxisVariableScores execute(VisualComponentList f1, int k);
+}
