+++ src/main/java/edu/uiuc/zenvisage/zql/functions/T.java
