+++ src/main/java/edu/uiuc/zenvisage/zql/functions/TIncreasingness.java
